各ファイルの記述修正及び確認
