+++ tsconfig.json
@@ -22,7 +22,7 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx"
-  ],
+, "src/next.config.js", "src/_document.js", "src/_app.js", "src/theme.js"  ],
   "exclude": [
     "node_modules"
   ]
